scotland-trip-planner/src/data/scraper.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Web scraping utilities for the Scotland Trip Planner.

This module provides functions for scraping data from websites,
particularly the Walk Highlands website for hiking route information.
"""

import requests
from bs4 import BeautifulSoup

class WebScraper:
    """Class for scraping data from websites."""
    
    def __init__(self, user_agent=None):
        """Initialize the WebScraper with the provided user agent."""
        self.user_agent = user_agent or "ScotlandTripPlanner/1.0"
        self.session = requests.Session()
        self.session.headers.update({"User-Agent": self.user_agent})
    
    def get_page(self, url):
        """Get the HTML content of a web page."""
        # TODO: Implement the scraping logic
        pass
    
    def extract_text(self, element, selector):
        """Extract text from an element using a CSS selector."""
        # TODO: Implement the extraction logic
        pass
    
    def extract_attribute(self, element, selector, attribute):
        """Extract an attribute from an element using a CSS selector."""
        # TODO: Implement the extraction logic
        pass