scotland-trip-planner/requirements.txt
# API and Web Requests
requests>=2.28.0

# Web Scraping
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Data Processing
pandas>=1.4.0
numpy>=1.22.0

# Geographical Calculations
geopy>=2.2.0

# Environment Variables
python-dotenv>=0.20.0

# Web Interface
Flask>=2.1.0
Jinja2>=3.1.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0